- Fix compiler warnings for MCPKit - largely typecast related  - Fix further compiler warnings, including a couple of bugs  - Disable the -wselector warnings (Multiple definition types for selector) as they're currently unsupported in LLVM  - Disable the -wconversion warnings (Prototype conversion) warnings as we usually can't affect protocol declaration  - Disable the -wstrict-selector-match (Strict selector matching) due to too many false positives

+++ Source/xibLocalizationPostprocessor.m
@@ -51,7 +51,7 @@ int main(int argc, const char *argv[])
                 }
             }
             
-            NSLog(@"Warning: skipped garbage input line %d, contents: \"%@\"", lineCount, line);
+            NSLog(@"Warning: skipped garbage input line %lu, contents: \"%@\"", (unsigned long)lineCount, line);
         }
         
         if (outputStrings.length && ![outputStrings writeToFile:[NSString stringWithUTF8String:argv[2]] atomically:NO encoding:usedEncoding error:&error]) {
